requirements.txt
asgiref==3.6.0
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.37.0
distlib==0.3.8
Django==4.2
django-cloudinary-storage==0.3.0
django-modeltranslation==0.18.9
filelock==3.13.1
idna==3.6
Pillow==10.1.0
platformdirs==4.1.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.25.0
